fix: move all to predicates_db
